+++ src/Presentation/Nop.Web/Nop.Web.csproj
@@ -279,6 +279,8 @@
     <Compile Include="Factories\BlogModelFactory.cs" />
     <Compile Include="Factories\IOrderModelFactory.cs" />
     <Compile Include="Factories\OrderModelFactory.cs" />
+    <Compile Include="Factories\IProfileModelFactory.cs" />
+    <Compile Include="Factories\ProfileModelFactory.cs" />
     <Compile Include="Global.asax.cs">
       <DependentUpon>Global.asax</DependentUpon>
     </Compile>
